Fix precision on playback preview.
Playback was unable to skip notes during the small window where it just hasn't reached the next row.

+++ src/Editor/notefield.cpp
@@ -591,7 +591,7 @@ void drawNotes()
 	const int signedScale = gView->hasReverseScroll() ? -scale : scale;
 	const int maxY = gView->getHeight() + 32;
 	const bool isPreview = !gMusic->isPaused() && gView->hasChartPreview();
-	const int currentRow = gView->getCursorBeat() * ROWS_PER_BEAT;
+	const double currentRow = gView->getCursorBeat() * ROWS_PER_BEAT;
 
 	auto noteskin = gNoteskin->get();
